requirements.txt
streamlit 
pytube 
ffmpeg-python